requirements.txt
PyPDF2~=3.0.1
openpyxl~=3.1.2
pillow~=10.3.0